requirements.txt
asgiref==3.8.1
Django==4.2.15
django-restframework==0.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
